Move trace-specific code into trace.go
